chore: label class modules with original sections
