+++ engine/__init__.py
@@ -1,0 +1,1 @@
+# Game engine module
