+++ mcts/__init__.py
@@ -1,0 +1,1 @@
+# Monte Carlo Tree Search components
