+++ self_play/__init__.py
@@ -1,0 +1,1 @@
+# self-play utilities
